Refactoring like Rubygems code.
