--- lib/uuid/uuid5.rb
@@ -1,12 +1,0 @@
-require 'digest/sha1'
-
-class UUID5 < NameBasedUUID
-
-  def version
-    5
-  end
-
-  def hash(target)
-    Digest::SHA1.hexdigest(target)[0, 32]
-  end
-end
