+++ lib/uuidgen.rb
@@ -1,0 +1,7 @@
+require 'uuidgen/version'
+require 'uuidgen/uuid'
+require 'uuidgen/name_space'
+require 'uuidgen/name_based_uuid'
+require 'uuidgen/uuid3'
+require 'uuidgen/uuid5'
+
