+++ lib/uuidgen/version.rb
@@ -1,0 +1,4 @@
+module UUIDGen
+  VERSION = "0.0.2"
+end
+
